Pacsea::ui_helpers

Function trigger_recent_preview

pub fn trigger_recent_preview(
    app: &AppState,
    preview_tx: &UnboundedSender<PackageItem>,
)

Trigger an asynchronous preview fetch for the currently selected Recent query, if applicable.

This function exits early unless:

The Recent pane has focus

There is a valid selection within bounds of the filtered view

The corresponding query string exists

When conditions are met, a Tokio task is spawned to resolve the query to a candidate package (via fetch_first_match_for_query) and send it over the provided preview_tx. Send errors are ignored to keep the UI responsive even if downstream receivers were dropped.